Conflicts:
	cocos2d-ios.xcodeproj/project.pbxproj
	cocos2d-tests-ios.xcodeproj/project.pbxproj
	cocos2d-ui-tests/ios/AppDelegate.m
	cocos2d/Platforms/iOS/CCAppDelegate.m
	cocos2d/Support/CCFileUtils.h

+++ cocos2d-ui/Platform/CCPlatformTextField.m
@@ -10,7 +10,7 @@
 
 @implementation CCPlatformTextField
 @synthesize delegate=_delegate;
-- (void) positionInControl:(CCControl *)control padding:(float)padding {
+- (void) positionInControl:(CCControl *)control padding:(CGFloat)padding {
     
 }
 - (void) onEnterTransitionDidFinish {
